Rename master/slave to controller/worker in the testing infrastructure.
This is just to avoid the controversial pair of words
"master/slave". The change aligns with how those things were named
in Scala Native.

The method `sbt.testing.Framework.slaveRunner` is not renamed,
because doing so would break binary compatibility.

In BinaryIncompatibilities, we merge `TestCommon` and
`TestAdapter`. `TestCommon` used to be reused for the `test-bridge`
artifact, but that one does not have any binary compatibility
constraints anymore, so the distinction was artificial.